requirements.txt
distlib==0.3.7
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
pipenv==2023.11.15
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1